Fixed another bug

+++ Inference/letter_recognizer_inference/train_letter_recognizer.py
@@ -28,7 +28,7 @@ class CSVLetterDataset(Dataset):
     def __init__(self, csv_path):
         self.data = pd.read_csv(csv_path, header=None).values
         self.labels = self.data[:, 0].astype(int)  - 1
-        self.images = self.data[:, 1:].reshape(-1, 28, 28)  
+        self.images = self.data[:, 1:].reshape(-1, 28, 28).astype('float32')  
         self.images = self.images.astype('float32') / 255.0
         self.images = torch.tensor(self.images).unsqueeze(1)  
         self.labels = torch.tensor(self.labels, dtype=torch.long)
